+++ lib/archives.js
@@ -12,7 +12,6 @@ module.exports = function(posts) {
     }
 
     posts.forEach(function(post, i) {
-        console.log(year(post.created))
 
         if (i == 0) {
