+++ lib/feed.js
@@ -22,7 +22,7 @@ module.exports = function(posts) {
             title: post.title,
             url: config.url + post.path,
             author: post.author.name,
-            date: post.update,
+            date: post.updated,
             description: post.body
         })
     })
